requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
numpy==1.24.3
opencv-python-headless==4.8.0.74
Pillow==9.5.0
gunicorn==20.1.0